Fix merge conflicts.

+++ modules/membership/includes/Admin/Services/Stripe/StripeService.php
@@ -308,10 +308,16 @@ class StripeService {
 			'type' => $payment_data['type'],
 		);
 
-		if ( isset( $payment_data['coupon'] ) && ! empty( $payment_data['coupon'] ) && ur_check_module_activation( 'coupon' ) ) {
-			$coupon_details  = ur_get_coupon_details( $payment_data['coupon'] );
-			$discount_amount = ( 'fixed' === $coupon_details['coupon_discount_type'] ) ? $coupon_details['coupon_discount'] : $amount * $coupon_details['coupon_discount'] / 100;
-			$amount          = $amount - $discount_amount;
+		if ( isset( $payment_data['upgrade'] ) && $payment_data['upgrade'] ) {
+			$amount = $payment_data['amount'];
+
+		} elseif ( isset( $payment_data['coupon'] ) && ! empty( $payment_data['coupon'] ) && ur_check_module_activation( 'coupon' ) ) {
+			$coupon_details = ur_get_coupon_details( $payment_data['coupon'] );
+
+			if ( isset( $coupon_details['coupon_discount_type'] ) && isset( $coupon_details['coupon_discount'] ) ) {
+				$discount_amount = ( 'fixed' === $coupon_details['coupon_discount_type'] ) ? $coupon_details['coupon_discount'] : $amount * $coupon_details['coupon_discount'] / 100;
+				$amount          = $amount - $discount_amount;
+			}
 		}
 
 		if ( 'JPY' === $currency ) {
@@ -403,7 +409,7 @@ class StripeService {
 					)
 				);
 
-				$intent                    = \Stripe\PaymentIntent::create(
+				$intent = \Stripe\PaymentIntent::create(
 					array(
 						'amount'               => $amount,
 						'currency'             => $currency,
@@ -411,6 +417,7 @@ class StripeService {
 						'customer'             => $customer->id,
 					)
 				);
+
 				$response['client_secret'] = $intent->client_secret;
 
 				PaymentGatewayLogging::log_transaction_success(
@@ -451,20 +458,33 @@ class StripeService {
 	}
 
 	public function update_order( $data ) {
-		$transaction_id = $data['payment_result']['paymentIntent']['id'] ?? '';
-		$payment_status = sanitize_text_field( $data['payment_status'] );
-		$member_id      = absint( $_POST['member_id'] );
-		$is_upgrading   = ur_string_to_bool( get_user_meta( $member_id, 'urm_is_upgrading', true ) );
+
+		$transaction_id         = $data['payment_result']['paymentIntent']['id'] ?? '';
+		$payment_status         = sanitize_text_field( $data['payment_status'] );
+		$member_id              = absint( $_POST['member_id'] );
+		$membership_process     = urm_get_membership_process( $member_id );
+		$selected_membership_id = isset( $_POST['selected_membership_id'] ) && '' !== $_POST['selected_membership_id'] ? absint( $_POST['selected_membership_id'] ) : 0;
+		$current_membership_id  = isset( $_POST['current_membership_id'] ) && '' !== $_POST['current_membership_id'] ? absint( $_POST['current_membership_id'] ) : 0;
+		$is_purchasing_multiple = ! empty( $membership_process['multiple'] ) && in_array( $selected_membership_id, $membership_process['multiple'] );
+		$is_upgrading           = ! empty( $membership_process['upgrade'] ) && isset( $membership_process['upgrade'][ $current_membership_id ] );
+		$transaction_id         = $data['payment_result']['paymentIntent']['id'] ?? '';
+
+		if ( empty( $transaction_id ) ) {
+			$transaction_id = $data['payment_result']['latest_invoice']['payment_intent']['next_action']['use_stripe_sdk']['directory_server_encryption']['server_transaction_id'] ?? '';
+		}
+
+		$three_d_secure_2_source = $data['payment_result']['latest_invoice']['payment_intent']['next_action']['use_stripe_sdk']['three_d_secure_2_source'] ?? '';
 
 		PaymentGatewayLogging::log_webhook_received(
 			'stripe',
 			'Stripe payment confirmation callback received',
 			array(
-				'webhook_type'   => 'payment_confirmation',
-				'transaction_id' => $transaction_id,
-				'payment_status' => $payment_status,
-				'member_id'      => $member_id,
-				'is_upgrade'     => $is_upgrading,
+				'webhook_type'           => 'payment_confirmation',
+				'transaction_id'         => $transaction_id,
+				'payment_status'         => $payment_status,
+				'member_id'              => $member_id,
+				'is_upgrade'             => $is_upgrading,
+				'is_purchasing_multiple' => $is_purchasing_multiple,
 			)
 		);
 
@@ -494,9 +514,10 @@ class StripeService {
 		$membership_type  = $membership_metas['type'] ?? 'unknown';
 
 		if ( 'failed' === $payment_status ) {
-			$is_renewing = ur_string_to_bool( get_user_meta( $member_id, 'urm_is_member_renewing', true ) );
-			$error_msg   = __( 'Stripe Payment failed.', 'user-registration' );
-			$error_msg   = $data['payment_result']['error']['message'] ?? $error_msg;
+			$is_renewing = ! empty( $membership_process['renew'] ) && in_array( $latest_order['item_id'], $membership_process['renew'] );
+
+			$error_msg = __( 'Stripe Payment failed.', 'user-registration' );
+			$error_msg = $data['payment_result']['error']['message'] ?? $error_msg;
 
 			PaymentGatewayLogging::log_transaction_failure(
 				'stripe',
@@ -510,12 +531,13 @@ class StripeService {
 				)
 			);
 
-			if ( ! $is_upgrading && ! $is_renewing ) {
+			if ( ! $is_upgrading && ! $is_renewing && ! $is_purchasing_multiple ) {
 				wp_delete_user( absint( $member_id ) );
 				$this->members_orders_repository->delete_member_order( $member_id );
 			}
 			if ( $is_renewing ) {
-				update_user_meta( $member_id, 'urm_is_member_renewing', false );
+				unset( $membership_process['upgrade'][ $latest_order['item_id'] ] );
+				update_user_meta( $member_id, 'urm_membership_process', $membership_process );
 			}
 
 			$response['message'] = $error_msg;
@@ -535,7 +557,7 @@ class StripeService {
 			$membership                     = $this->membership_repository->get_single_membership_by_ID( $member_order['item_id'] );
 			$membership_metas               = wp_unslash( json_decode( $membership['meta_value'], true ) );
 			$membership_metas['post_title'] = $membership['post_title'];
-			$member_subscription            = $this->members_subscription_repository->get_member_subscription( $member_id );
+			$member_subscription            = $this->members_subscription_repository->get_subscription_data_by_member_and_membership_id( $member_id, $member_order['item_id'] );
 			$is_order_updated               = $this->members_orders_repository->update(
 				$member_order['ID'],
 				array(
@@ -544,7 +566,7 @@ class StripeService {
 				)
 			);
 
-			if ( $is_order_updated && 'paid' === $member_order['order_type'] ) {
+			if ( ( $is_order_updated && 'paid' === $member_order['order_type'] ) || ( $is_order_updated && ! empty( $three_d_secure_2_source ) && 'subscription' === $member_order['order_type'] ) ) {
 				$this->members_subscription_repository->update(
 					$member_subscription['ID'],
 					array(
@@ -600,6 +622,7 @@ class StripeService {
 					)
 				);
 			}
+
 			$response = $this->sendEmail( $member_order['ID'], $member_subscription, $membership_metas, $member_id, $response );
 		}
 
@@ -613,10 +636,14 @@ class StripeService {
 		$membership_metas = wp_unslash( json_decode( $membership['meta_value'], true ) );
 
 		$membership_metas['post_title'] = $membership['post_title'];
-		$member_subscription            = $this->members_subscription_repository->get_member_subscription( $member_id );
-		$is_automatic                   = 'automatic' === get_option( 'user_registration_renewal_behaviour', 'automatic' );
-		$is_renewing                    = ur_string_to_bool( get_user_meta( $member_id, 'urm_is_member_renewing', true ) );
-		$membership_type                = $membership_metas['type'] ?? 'unknown';
+
+		$member_subscription = $this->members_subscription_repository->get_subscription_data_by_member_and_membership_id( $member_id, $membership['ID'] );
+		$is_automatic        = 'automatic' === get_option( 'user_registration_renewal_behaviour', 'automatic' );
+
+		$membership_process = urm_get_membership_process( $member_id );
+		$is_renewing        = ! empty( $membership_process['renew'] ) && in_array( $member_order['item_id'], $membership_process['renew'] );
+
+		$membership_type = $membership_metas['type'] ?? 'unknown';
 
 		PaymentGatewayLogging::log_api_request(
 			'stripe',
@@ -702,6 +729,7 @@ class StripeService {
 
 			// handle coupon section
 			$order_detail = $this->orders_repository->get_order_detail( $member_order['ID'] );
+
 			if ( ! empty( $order_detail['coupon'] ) ) {
 				$coupon_details = ur_get_coupon_details( $order_detail['coupon'] );
 				if ( ! empty( $coupon_details['stripe_coupon_id'] ) ) {
@@ -735,9 +763,16 @@ class StripeService {
 							$first_month_price = $new_price - $current_price;
 
 							if ( $new_price > $current_price ) {
-								$discount_amount = $new_price - $first_month_price;
-								$currency        = get_option( 'user_registration_payment_currency', 'USD' );
-								$amount          = ( 'JPY' === $currency ) ? $discount_amount : $discount_amount * 100;
+								if ( isset( $order_detail['coupon'] ) && ! empty( $order_detail['coupon'] ) && ur_check_module_activation( 'coupon' ) ) {
+									$coupon_details  = ur_get_coupon_details( $order_detail['coupon'] );
+									$discount_amount = ( 'fixed' === $coupon_details['coupon_discount_type'] ) ? $coupon_details['coupon_discount'] : $first_month_price * $coupon_details['coupon_discount'] / 100;
+									$amount          = $new_price - $discount_amount;
+								} else {
+									$amount = $new_price - $first_month_price;
+								}
+
+								$currency = get_option( 'user_registration_payment_currency', 'USD' );
+								$amount   = ( 'JPY' === $currency ) ? $amount : $amount * 100;
 
 								PaymentGatewayLogging::log_general(
 									'stripe',
@@ -748,7 +783,7 @@ class StripeService {
 										'member_id'       => $member_id,
 										'old_price'       => $current_price,
 										'new_price'       => $new_price,
-										'discount_amount' => $discount_amount,
+										'discount_amount' => $amount,
 									)
 								);
 
@@ -847,17 +882,23 @@ class StripeService {
 				)
 			);
 
-			if ( 'active' === $subscription_status || 'trialing' === $subscription_status ) {
+			$three_ds2_source = $subscription->latest_invoice->payment_intent->next_action->use_stripe_sdk->three_d_secure_2_source ?? '';
+
+			if ( 'active' === $subscription_status || 'trialing' === $subscription_status || ( 'incomplete' === $subscription_status && ! empty( $three_ds2_source ) ) ) {
+				$status = ( 'incomplete' === $subscription_status && ! empty( $three_ds2_source ) ) ? 'pending' : 'completed';
 				$this->members_orders_repository->update(
 					$member_order['ID'],
 					array(
-						'status'         => 'completed',
+						'status'         => $status,
 						'transaction_id' => $subscription->id,
 					)
 				);
 				switch ( $subscription_status ) {
 					case 'trialing':
 						$subscription_status = ( $is_upgrading && ! empty( $next_subscription['delayed_until'] ) ) ? 'active' : 'trial';
+						break;
+					case ( 'incomplete' === $subscription_status && ! empty( $three_ds2_source ) ):
+						$subscription_status = 'pending';
 						break;
 					default:
 						break;
@@ -950,6 +991,7 @@ class StripeService {
 			$order_detail['coupon_discount_type'] = get_user_meta( $order_detail['user_id'], 'ur_coupon_discount_type', true );
 
 		}
+
 		$email_data = array(
 			'subscription'     => $member_subscription,
 			'order'            => $order_detail,
@@ -999,7 +1041,7 @@ class StripeService {
 			}
 		}
 		// delete coupon if new changes introduced
-		if ( $coupon_exists && ( $old_coupon_data['coupon_discount'] !== $data['post_meta_data']['coupon_discount'] || $old_coupon_data['coupon_discount_type'] !== $data['post_meta_data']['post_meta_data'] ) ) {
+		if ( $coupon_exists && ( $old_coupon_data['coupon_discount'] !== $data['post_meta_data']['coupon_discount'] || $old_coupon_data['coupon_discount_type'] !== $data['post_meta_data']['coupon_discount_type'] ) ) {
 			$coupon = \Stripe\Coupon::retrieve( $stripe_coupon_id );
 			$coupon->delete();
 		}
@@ -1144,15 +1186,15 @@ class StripeService {
 	}
 
 	public function handle_webhook( $event, $subscription_id ) {
-        // Verify that the event was sent by Stripe
-		if( isset( $event[ 'id' ] ) ) {
-            try {
-				$event_id = sanitize_text_field( $event[ 'id' ] );
-                $event = (array)\Stripe\Event::retrieve( $event_id );
-            } catch( \Exception $e ) {
+		// Verify that the event was sent by Stripe
+		if ( isset( $event['id'] ) ) {
+			try {
+				$event_id = sanitize_text_field( $event['id'] );
+				$event    = (array) \Stripe\Event::retrieve( $event_id );
+			} catch ( \Exception $e ) {
 				die();
-            }
-        } else {
+			}
+		} else {
 			die();
 		}
 		switch ( $event['type'] ) {
@@ -1262,7 +1304,8 @@ class StripeService {
 				'status'            => $subscription_status,
 			)
 		);
-		$is_renewing = ur_string_to_bool( get_user_meta( $member_id, 'urm_is_member_renewing', true ) );
+		$membership_process = urm_get_membership_process( $member_id );
+		$is_renewing        = ! empty( $membership_process['renew'] ) && in_array( $membership_id, $membership_process['renew'] );
 
 		if ( $is_renewing ) {
 			$subscription_service = new SubscriptionService();
@@ -1402,8 +1445,12 @@ class StripeService {
 		$refund_response = $this->refund( $last_order, $subscription );
 
 		delete_user_meta( $member_id, 'urm_previous_subscription_data' );
-		delete_user_meta( $member_id, 'urm_is_upgrading' );
-		delete_user_meta( $member_id, 'urm_is_upgrading_to' );
+		$membership_process = urm_get_membership_process( $member_id );
+
+		if ( ! empty( $membership_process ) && isset( $membership_process['upgrade'][ $_POST['current_membership_id'] ] ) ) {
+			unset( $membership_process['upgrade'][ $_POST['current_membership_id'] ] );
+			update_user_meta( $member_id, 'urm_membership_process', $membership_process );
+		}
 	}
 
 	/**
